# openapi.json
{"openapi":"3.1.0","info":{"title":"rust-svelte-todo","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/auth/me":{"get":{"tags":["Auth"],"summary":"Get current user","description":"Returns the profile of the currently authenticated user.\n\nAuthentication is accepted via `auth_token` cookie (preferred) or\n`Authorization: Bearer <token>` header.","operationId":"me","responses":{"200":{"description":"Current user profile","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/auth/signin":{"post":{"tags":["Auth"],"summary":"Sign in","description":"Authenticates an existing user and sets an `auth_token` HttpOnly cookie.\n\nThe cookie is used automatically for subsequent requests. The token is also returned\nin the response body for clients that prefer the `Authorization: Bearer` header.","operationId":"signin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninUser"}}},"required":true},"responses":{"200":{"description":"Authenticated, returns JWT token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"401":{"description":"Invalid email or password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/auth/signout":{"post":{"tags":["Auth"],"summary":"Sign out","description":"Clears the `auth_token` cookie.","operationId":"signout","responses":{"204":{"description":"Signed out, cookie cleared"}}}},"/todos":{"get":{"tags":["Todos"],"summary":"List all todos","description":"Returns all todos in the database.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"list_todos","parameters":[{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"List of todos","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Todo"}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Todos"],"summary":"Create a new todo","description":"Creates a new todo with a unique slug generated from the title.\n\nIf a slug collision occurs, it will attempt to generate a new slug up to 3 times\nbefore returning a conflict error.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"create_todo","parameters":[{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodo"}}},"required":true},"responses":{"201":{"description":"Todo created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Slug already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/todos/{id}":{"get":{"tags":["Todos"],"summary":"Get a todo by ID","description":"Returns a single todo identified by its ID.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"get_todo","parameters":[{"name":"id","in":"path","description":"The todo ID","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Todo found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Todo not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["Todos"],"summary":"Update a todo","description":"Updates the title, description, and completed status of an existing todo.\nThe slug is regenerated from the new title. If a slug collision occurs, it will\nattempt to generate a new slug up to 3 times before returning a conflict error.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"update_todo","parameters":[{"name":"id","in":"path","description":"The todo ID","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodo"}}},"required":true},"responses":{"200":{"description":"Todo updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Todo not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Slug already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Todos"],"summary":"Delete todo","description":"Deletes a todo identified by its ID. Returns 204 No Content on success.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"delete_todo","parameters":[{"name":"id","in":"path","description":"The todo ID","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Todo deleted"},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Todo not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/users":{"get":{"tags":["Users"],"summary":"List all users","description":"Returns all users in the database.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"list_users","parameters":[{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"List of users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Users"],"summary":"Create a new user","description":"Creates a new user with a unique slug generated from the full name.\n\nIf a slug collision occurs, it will attempt to generate a new slug up to 3 times\nbefore returning a conflict error.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"create_user","parameters":[{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUser"}}},"required":true},"responses":{"201":{"description":"User created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Slug already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/users/{id}":{"get":{"tags":["Users"],"summary":"Get a user by ID","description":"Returns a single user identified by its ID.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"get_user","parameters":[{"name":"id","in":"path","description":"The user ID","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["Users"],"summary":"Update a user","description":"Updates the full name, email, and role of an existing user.\nThe slug is regenerated from the new full name. If a slug collision occurs, it will\nattempt to generate a new slug up to 3 times before returning a conflict error.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"update_user","parameters":[{"name":"id","in":"path","description":"The user ID","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUser"}}},"required":true},"responses":{"200":{"description":"User updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Slug already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Users"],"summary":"Delete user","description":"Deletes a user identified by its ID. Returns 204 No Content on success.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"delete_user","parameters":[{"name":"id","in":"path","description":"The user ID","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"User deleted"},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"CreateTodo":{"type":"object","required":["title","description"],"properties":{"description":{"type":"string"},"title":{"type":"string"}}},"CreateUser":{"type":"object","required":["full_name","email","password","role"],"properties":{"email":{"type":"string"},"full_name":{"type":"string"},"password":{"type":"string"},"role":{"$ref":"#/components/schemas/UserRole"}}},"ErrorResponse":{"type":"object","required":["error"],"properties":{"error":{"type":"string"}}},"SigninUser":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string"},"password":{"type":"string"}}},"Todo":{"type":"object","required":["id","slug","title","description","completed"],"properties":{"completed":{"type":"boolean"},"description":{"type":"string"},"id":{"type":"string"},"slug":{"type":"string"},"title":{"type":"string"}}},"Token":{"type":"object","required":["token"],"properties":{"token":{"type":"string"}}},"UpdateTodo":{"type":"object","required":["title","description","completed"],"properties":{"completed":{"type":"boolean"},"description":{"type":"string"},"title":{"type":"string"}}},"UpdateUser":{"type":"object","required":["full_name","role"],"properties":{"full_name":{"type":"string"},"role":{"$ref":"#/components/schemas/UserRole"}}},"User":{"type":"object","required":["id","slug","full_name","email","role"],"properties":{"email":{"type":"string"},"full_name":{"type":"string"},"id":{"type":"string"},"role":{"$ref":"#/components/schemas/UserRole"},"slug":{"type":"string"}}},"UserRole":{"type":"string","enum":["admin","editor","viewer"]}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"tags":[{"name":"Todos"},{"name":"Auth"},{"name":"Users"}]}